Add custom image render delegates.
I first implemented a render delegate that uses css background image properties to center the passed image. However the performance on iPad when flicking was very poor. I back to using the canvas based renderer to fix which seems more performant but stretches the image (ugly). Investigating how to make the canvas preserve the aspect ratio of images.